+++ wp-content/themes/patus/functions.php
@@ -33,6 +33,8 @@ function patus_setup() {
 	// Add default posts and comments RSS feed links to head.
 	add_theme_support( 'automatic-feed-links' );
 
+	add_theme_support( 'title-tag' );
+
 	/*
 	 * Enable support for Post Thumbnails on posts and pages.
 	 *
@@ -62,6 +64,18 @@ function patus_setup() {
 		'default-color' => 'ffffff',
 		'default-image' => '',
 	) ) );
+
+	/**
+	 * Add support for core custom logo.
+	 *
+	 * @link https://codex.wordpress.org/Theme_Logo
+	 */
+	add_theme_support( 'custom-logo', array(
+		'height'      => 65,
+		'width'       => 180,
+		'flex-width'  => true,
+		'flex-height' => true,
+	) );
 }
 endif; // patus_setup
 add_action( 'after_setup_theme', 'patus_setup' );
@@ -78,8 +92,8 @@ function patus_widgets_init() {
 		'description'   => '',
 		'before_widget' => '<aside id="%1$s" class="widget %2$s">',
 		'after_widget'  => '</aside>',
-		'before_title'  => '<h1 class="widget-title">',
-		'after_title'   => '</h1>',
+		'before_title'  => '<h3 class="widget-title">',
+		'after_title'   => '</h3>',
 	) );
 }
 add_action( 'widgets_init', 'patus_widgets_init' );
@@ -125,7 +139,7 @@ function patus_scripts() {
                 input[type=\"submit\"] {
                     background: #{$primary};
                 }
-                #site-navigation.main-navigation.toggled .menu-toggle span,
+                #site-navigation.main-navigation .menu-toggle span,
                 #site-navigation .menu ul li a:hover,
                 #site-navigation .menu ul li a:focus,
                 #site-navigation .menu ul ul li a:hover,
@@ -163,6 +177,11 @@ function patus_scripts() {
                 .entry-meta {
                 	background: #{$secondary};
                 }";
+
+		if ( get_header_image() ) :
+			$custom_css .= '.site-header {  background-image: url('. esc_url( get_header_image() ) .'); background-repeat: no-repeat; background-size: cover; }';
+		endif;
+
     wp_add_inline_style( 'patus-style', $custom_css );
 
 	// Enqueue jQuery
@@ -208,3 +227,44 @@ require get_template_directory() . '/inc/customizer.php';
  * Add theme info page
  */
 require get_template_directory() . '/inc/dashboard.php';
+
+
+/**
+ * Include the TGM_Plugin_Activation class.
+ */
+require_once get_template_directory() . '/inc/tgm.php';
+add_action( 'tgmpa_register', 'patus_register_required_plugins' );
+function patus_register_required_plugins() {
+	/*
+	 * Array of plugin arrays. Required keys are name and slug.
+	 * If the source is NOT from the .org repo, then source is also required.
+	 */
+	$plugins = array(
+		// This is an example of how to include a plugin from the WordPress Plugin Repository.
+		array(
+			'name'      => 'Mega Menu plugin for WordPress',
+			'slug'      => 'easymega',
+			'required'  => false,
+		),
+	);
+	/*
+	 * Array of configuration settings. Amend each line as needed.
+	 *
+	 * TGMPA will start providing localized text strings soon. If you already have translations of our standard
+	 * strings available, please help us make TGMPA even better by giving us access to these translations or by
+	 * sending in a pull-request with .po file(s) with the translations.
+	 *
+	 * Only uncomment the strings in the config array if you want to customize the strings.
+	 */
+	$config = array(
+		'id'           => 'patus',                 // Unique ID for hashing notices for multiple instances of TGMPA.
+		'default_path' => '',                      // Default absolute path to bundled plugins.
+		'menu'         => 'tgmpa-install-plugins', // Menu slug.
+		'has_notices'  => true,                    // Show admin notices or not.
+		'dismissable'  => true,                    // If false, a user cannot dismiss the nag message.
+		'dismiss_msg'  => '',                      // If 'dismissable' is false, this message will be output at top of nag.
+		'is_automatic' => false,                   // Automatically activate plugins after installation or not.
+		'message'      => '',                      // Message to output right before the plugins table.
+	);
+	tgmpa( $plugins, $config );
+}