+++ wp-content/themes/patus/inc/template-tags.php
@@ -72,7 +72,7 @@ function patus_posted_on() {
 	}
 
 	if ( ! is_single() && ! post_password_required() && ( comments_open() || get_comments_number() ) ) {
-		echo '<span class="comments-link">';
+		echo ' <span class="comments-link">';
 		comments_popup_link( esc_html__( '0 Comment', 'patus' ), esc_html__( '1 Comment', 'patus' ), esc_html__( '% Comments', 'patus' ) );
 		echo '</span>';
 	}